server/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0
PyJWT>=2.8
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
gunicorn>=21.0
